keith_py/requirements.txt
cryptography~=3.4.7
pynput~=1.7.3
Pillow~=8.2.0
requests~=2.25.1
browserhistory~=0.1.2
opencv-python~=4.5.2.52